REFACTOR: refactoring TodoState, MainScreen
